requirements.txt
tensorflow>=2.15.0
gym-retro>=0.8.0
numpy>=1.24.0
opencv-python>=4.8.0 